requirements.txt
langchain
langchain-openai
python-dotenv
pdfplumber
pandas